Renamed EmailService to EmailServiceImpl and implemented EmailService interface. Implemented createAndSendRegistrationConfirmationEmail and createRegistrationConfirmationMessage methods
